sand & wanter simulation

+++ src/game/constants.rs
@@ -1,3 +1,3 @@
 ﻿pub const CHUNK_SIZE: usize = 64;
 pub const WORLD_CHUNK_SIZE: f32 = 300.0;
-pub const PIXEL_SIMULATION_TIMESTEP: f64 = 1.0 / 60.0;
+pub const PIXEL_SIMULATION_TIMESTEP: f32 = 1.0 / 60.0;
